5/31/2026

+++ app/Http/Controllers/User/Commissioning/FormController.php
@@ -837,7 +837,7 @@ class FormController extends Controller
 
     private function authorizeSubmission(CommissioningFormSubmission $submission): void
     {
-        abort_unless($submission->user_id === auth()->id() || auth()->user()?->isAdmin(), 403);
+        abort_unless((int) $submission->user_id === (int) auth()->id() || auth()->user()?->isAdmin(), 403);
     }
 
     private function attachmentStoragePath(CommissioningFormSubmissionAttachment $attachment): ?string
