+++ app/Http/Controllers/User/Qc/FormController.php
@@ -1419,7 +1419,7 @@ class FormController extends Controller
 
     private function authorizeSubmission(QcFormSubmission $submission): void
     {
-        abort_unless($submission->user_id === auth()->id() || auth()->user()?->isAdmin(), 403);
+        abort_unless((int) $submission->user_id === (int) auth()->id() || auth()->user()?->isAdmin(), 403);
     }
 
     private function attachmentStoragePath(QcFormSubmissionAttachment $attachment): ?string
